+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 # Unreleased
 
-- Use a variable to hold the Arguments
+- Use a variable to hold the Arguments.
+- Create a function to return a greeting message dependant on the hour of the day.
 
 ## v0.0.1 2020-08-18